More reasonable data types

+++ internal/http_app/api/chatroom.go
@@ -39,7 +39,7 @@ func GetChatroomAvatarNameByChatRoomID(writer http.ResponseWriter, request *http
 	user_id := formData["uid"].(int)
 	chatroom_id := formData["chatroom_id"].(string)
 	chatroom_type := "1"
-	if res, err := ChatRoomService.GetChatroomBaseInfo(chatroom_id, chatroom_type, uint(user_id)); err != nil {
+	if res, err := ChatRoomService.GetChatroomBaseInfo(chatroom_id, chatroom_type, int64(user_id)); err != nil {
 		util.RespFail(writer, err.Error())
 	} else {
 		util.RespOk(writer, res, "")
@@ -63,10 +63,10 @@ func CreateGroup(writer http.ResponseWriter, request *http.Request) {
 	json.NewDecoder(request.Body).Decode(&formData)
 
 	var (
-		group_id uint
+		group_id int64
 		err      error
 	)
-	if group_id, err = ChatRoomService.CreateGroup(uint(formData["uid"].(int)), model.Group{
+	if group_id, err = ChatRoomService.CreateGroup(int64(formData["uid"].(int)), model.Group{
 		Name:   formData["name"].(string),
 		Avatar: formData["avatar"].(string),
 	}); err != nil {
@@ -86,7 +86,7 @@ func AddGroup(writer http.ResponseWriter, request *http.Request) {
 	// 调用json包的解析，解析请求body
 	json.NewDecoder(request.Body).Decode(&formData)
 
-	if ret, err := ChatRoomService.JoinGroup(uint(formData["uid"].(int)), formData["group_id"].(string), formData["remark"].(string)); err != nil {
+	if ret, err := ChatRoomService.JoinGroup(int64(formData["uid"].(int)), formData["group_id"].(string), formData["remark"].(string)); err != nil {
 		util.RespFail(writer, err.Error())
 	} else {
 		util.RespOk(writer, ret, "")
